+++ Specs/Scene/CameraSpec.js
@@ -2797,13 +2797,14 @@ describe('Scene/Camera', function() {
         var context = scene.context;
         var drawingBufferWidth = context.drawingBufferWidth;
         var drawingBufferHeight = context.drawingBufferHeight;
+        var pixelRatio = 1.0;
 
         // Compute expected pixel size
         var distance = camera.distanceToBoundingSphere(sphere);
-        var pixelDimensions = camera.frustum.getPixelDimensions(drawingBufferWidth, drawingBufferHeight, distance, new Cartesian2());
+        var pixelDimensions = camera.frustum.getPixelDimensions(drawingBufferWidth, drawingBufferHeight, pixelRatio, distance, new Cartesian2());
         var expectedPixelSize = Math.max(pixelDimensions.x, pixelDimensions.y);
 
-        var pixelSize = camera.getPixelSize(sphere, drawingBufferWidth, drawingBufferHeight);
+        var pixelSize = camera.getPixelSize(sphere, drawingBufferWidth, drawingBufferHeight, pixelRatio);
         expect(pixelSize).toEqual(expectedPixelSize);
     });
 
@@ -2830,6 +2831,15 @@ describe('Scene/Camera', function() {
 
         expect(function() {
             camera.getPixelSize(sphere, 10);
+        }).toThrowDeveloperError();
+    });
+
+    it('getPixelSize throws when there is no pixel ratio', function() {
+        scene.mode = SceneMode.SCENE3D;
+        var sphere = new BoundingSphere(Cartesian3.ZERO, 0.5);
+
+        expect(function() {
+            camera.getPixelSize(sphere, 1.0, 1.0);
         }).toThrowDeveloperError();
     });